don't delete import/ and export/ for no good reason

+++ repo-overlay.pl
@@ -407,7 +407,7 @@ if ($do_new_symlinks or !defined($apply_repo)) {
     nsystem("rm -rf $outdir/export/*");
     nsystem("rm -rf $outdir/import/.repo");
     nsystem("rm -rf $outdir/export/.repo");
-} else {
+} elsif (defined($apply_repo)) {
     # rm -rf dangling-symlink/ doesn't delete anything. Learn
     # something new every day.
     nsystem("rm -rf $outdir/import/" . ($apply_repo =~ s/\/$//r));
